api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
firebase-admin==6.6.0
google-cloud-firestore>=2.17,<3
google-cloud-storage>=2.17,<3
pydantic==2.10.4
requests>=2.32.3,<3